% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawSemHints-class.R
\name{buildDrawSemHints}
\alias{buildDrawSemHints}
\title{Build drawSemHints for a Model}
\usage{
buildDrawSemHints(schema, model_id)
}
\arguments{
\item{schema}{The full schema list}

\item{model_id}{Which model's fragment to escrow}
}
\value{
A \code{drawSemHints} object
}
\description{
Constructs the \code{drawSemHints} object \code{as.MxModel()} attaches to the built
\code{mxModel} at \verb{@options$drawSemHints}.
}
\keyword{internal}
