% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{buildMxData}
\alias{buildMxData}
\title{Build mxData Object}
\usage{
buildMxData(model, data)
}
\arguments{
\item{model}{The model list from schema}

\item{data}{A named list of data.frames}
}
\value{
An mxData object (for single dataset) or list of mxData objects
}
\description{
Constructs an mxData object from schema dataset node(s) and data.
}
\keyword{internal}
