% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphModel-methods.R
\name{builtModel<-}
\alias{builtModel<-}
\title{Set Cached Built Model}
\usage{
builtModel(object) <- value
}
\arguments{
\item{object}{A GraphModel object}

\item{value}{A model object}
}
\value{
The modified GraphModel (invisibly)
}
\description{
Set Cached Built Model
}
\keyword{internal}
