% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphModel-methods.R
\name{builtModel}
\alias{builtModel}
\title{Get Cached Built Model}
\usage{
builtModel(object)
}
\arguments{
\item{object}{A GraphModel object}
}
\value{
The lastBuiltModel from the GraphModel (or NULL)
}
\description{
Get Cached Built Model
}
\keyword{internal}
