% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphModel-methods.R
\name{data<-}
\alias{data<-}
\title{Set Data in GraphModel}
\usage{
data(object) <- value
}
\arguments{
\item{object}{A GraphModel object}

\item{value}{A new data list}
}
\value{
The modified GraphModel (invisibly)
}
\description{
Set Data in GraphModel
}
\examples{
\dontrun{
g <- as.GraphModel(schema_list)
data(g) <- list(newdata = new_df)
}

}
