% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphModel-methods.R
\name{data}
\alias{data}
\title{Extract Data from GraphModel}
\usage{
data(object)
}
\arguments{
\item{object}{A GraphModel object}
}
\value{
The data list from the GraphModel
}
\description{
Extract Data from GraphModel
}
\examples{
\dontrun{
g <- as.GraphModel(schema_list, data = list(df1 = my_df))
d <- data(g)
}

}
