% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphModel-methods.R
\name{$,GraphModel-method}
\alias{$,GraphModel-method}
\alias{$<-,GraphModel-method}
\title{Access GraphModel Slots with $ Operator}
\usage{
\S4method{$}{GraphModel}(x, name)

\S4method{$}{GraphModel}(x, name) <- value
}
\arguments{
\item{x}{A GraphModel object}

\item{name}{The name of the slot to replace}

\item{value}{The new value for the slot}
}
\value{
The contents of the named slot

The modified GraphModel (invisibly)
}
\description{
Extract or replace slots in a GraphModel object using the $ operator.
Provides convenient access to schema, data, metadata, and lastBuiltModel.

Replace slots in a GraphModel object using the $<- operator.
}
\examples{
\dontrun{
gm <- as.GraphModel(schema, data = list(mydata = df))
gm$schema       # Access schema
gm$data$mydata  # Access data
gm$metadata     # Access metadata
}

\dontrun{
gm <- as.GraphModel(schema)
gm$data <- list(mydata = df)
gm$metadata$note <- "Updated"
}

}
