% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getConstantNodeLabel}
\alias{getConstantNodeLabel}
\title{Get Constant Node Label}
\usage{
getConstantNodeLabel(nodes)
}
\arguments{
\item{nodes}{List of node specifications}
}
\value{
The label of the constant node (or NULL if none found)
}
\description{
Finds the constant node in a list of nodes and returns its label.
}
\keyword{internal}
