% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphModel-methods.R
\docType{methods}
\name{GraphModel-methods}
\alias{GraphModel-methods}
\alias{show,GraphModel-method}
\title{GraphModel Methods: Accessors and Display}
\usage{
\S4method{show}{GraphModel}(object)
}
\description{
Methods for inspecting and modifying GraphModel objects.
}
