% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{inferLatentVariables}
\alias{inferLatentVariables}
\title{Infer Latent Variables from Nodes and Manifest Variables}
\usage{
inferLatentVariables(nodes, manifest_vars)
}
\arguments{
\item{nodes}{List of node specifications}

\item{manifest_vars}{Character vector of manifest variable labels}
}
\value{
Character vector of latent variable labels
}
\description{
A variable is latent if it's not manifest.
}
\keyword{internal}
