% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{inferManifestVariables}
\alias{inferManifestVariables}
\title{Infer Manifest Variables from Nodes and Paths}
\usage{
inferManifestVariables(nodes, paths)
}
\arguments{
\item{nodes}{List of node specifications}

\item{paths}{List of path specifications}
}
\value{
Character vector of manifest variable labels
}
\description{
A variable is manifest if it has variableCharacteristics$manifestLatent="manifest"
or (if not explicitly set) has incoming paths from dataset nodes.
}
\keyword{internal}
