% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphModel-methods.R
\name{metadata<-}
\alias{metadata<-}
\title{Set Metadata in GraphModel}
\usage{
metadata(object) <- value
}
\arguments{
\item{object}{A GraphModel object}

\item{value}{A new metadata list}
}
\value{
The modified GraphModel (invisibly)
}
\description{
Set Metadata in GraphModel
}
\examples{
\dontrun{
g <- as.GraphModel(schema_list)
metadata(g) <- list(positions = list(...))
}

}
