% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphModel-methods.R
\name{metadata}
\alias{metadata}
\title{Extract Metadata from GraphModel}
\usage{
metadata(object)
}
\arguments{
\item{object}{A GraphModel object}
}
\value{
The metadata list from the GraphModel
}
\description{
Extract Metadata from GraphModel
}
\examples{
\dontrun{
g <- as.GraphModel(schema_list)
m <- metadata(g)
}

}
