% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{removeDataPath}
\alias{removeDataPath}
\title{Remove a Data Path from a GraphModel}
\usage{
removeDataPath(graphModel, from, to)
}
\arguments{
\item{graphModel}{A \code{GraphModel} object to modify.}

\item{from}{Character. Label of the dataset node.}

\item{to}{Character. Label of the variable node.}
}
\value{
The modified \code{graphModel} object (invisibly).
}
\description{
Remove a Data Path from a GraphModel
}
