% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphTool.R
\name{renderGraphTool}
\alias{renderGraphTool}
\title{Shiny output binding for graphTool}
\usage{
renderGraphTool(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Shiny output binding for graphTool
}
\keyword{internal}
