% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphModel-methods.R
\name{schema<-}
\alias{schema<-}
\title{Set Schema in GraphModel}
\usage{
schema(object) <- value
}
\arguments{
\item{object}{A GraphModel object}

\item{value}{A new schema list}
}
\value{
The modified GraphModel (invisibly)
}
\description{
Set Schema in GraphModel
}
\examples{
\dontrun{
g <- as.GraphModel(schema_list)
schema(g) <- new_schema
}

}
