% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphModel-methods.R
\name{schema}
\alias{schema}
\title{Extract Schema from GraphModel}
\usage{
schema(object)
}
\arguments{
\item{object}{A GraphModel object}
}
\value{
The schema list from the GraphModel
}
\description{
Extract Schema from GraphModel
}
\examples{
\dontrun{
g <- as.GraphModel(schema_list)
s <- schema(g)
}

}
