% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{storeOptimizationMetadata}
\alias{storeOptimizationMetadata}
\title{Store Optimization Metadata}
\usage{
storeOptimizationMetadata(paths)
}
\arguments{
\item{paths}{List of path specifications}
}
\value{
A list of parameter metadata
}
\description{
Extracts bounds and priors from paths for later application.
}
\keyword{internal}
