% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{buildPathList}
\alias{buildPathList}
\title{Build Path List for mxModel}
\usage{
buildPathList(paths, constantNodeLabel = NULL)
}
\arguments{
\item{paths}{List of path specifications}

\item{constantNodeLabel}{Label of the constant node (if any)}
}
\value{
A list of path specifications, ready for mxPath()
}
\description{
Converts schema paths to a list of mxPath specifications.
}
\keyword{internal}
