% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{loadSchema}
\alias{loadSchema}
\title{Load a Graph Schema from JSON File}
\usage{
loadSchema(filepath)
}
\arguments{
\item{filepath}{Character. Path to JSON file containing schema.}
}
\value{
A list representing the schema (JSON structure).
}
\description{
Load a Graph Schema from JSON File
}
\details{
The schema defines the structure of a structural equation model:
nodes, paths, constraints, data sources, etc.
}
\examples{
\dontrun{
schema <- loadSchema("model.json")
validateSchema(schema)
}

}
