% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{removeDataPath}
\alias{removeDataPath}
\title{Remove a Data Path from a GraphModel}
\usage{
removeDataPath(graphModel, from, to)
}
\arguments{
\item{graphModel}{A \code{GraphModel} object to modify.}

\item{from}{Character. Label of the dataset node.}

\item{to}{Character. Label of the variable node.}
}
\value{
The modified \code{graphModel} object (invisibly).
}
\description{
Remove a Data Path from a GraphModel
}
\seealso{
Other GraphModel verbs:
\code{\link[=addConstant]{addConstant()}},
\code{\link[=addDataPath]{addDataPath()}},
\code{\link[=addDataset]{addDataset()}},
\code{\link[=addPath]{addPath()}},
\code{\link[=addVariable]{addVariable()}},
\code{\link[=changePath]{changePath()}},
\code{\link[=convertPath]{convertPath()}},
\code{\link[=nodes]{nodes()}},
\code{\link[=paths]{paths()}},
\code{\link[=removeConstant]{removeConstant()}},
\code{\link[=removeDataset]{removeDataset()}},
\code{\link[=removePath]{removePath()}},
\code{\link[=removeVariable]{removeVariable()}}
}
\concept{GraphModel verbs}
