% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{removeDataset}
\alias{removeDataset}
\title{Remove a Dataset Node from a GraphModel}
\usage{
removeDataset(graphModel, label)
}
\arguments{
\item{graphModel}{A \code{GraphModel} object to modify.}

\item{label}{Character. The label of the dataset node to remove.}
}
\value{
The modified \code{graphModel} object (invisibly).
}
\description{
Removing a dataset node also removes its \code{type: "data"} paths, and the
corresponding entries in \verb{@data} and \verb{@dataConnections}.
}
\seealso{
Other GraphModel verbs:
\code{\link[=addConstant]{addConstant()}},
\code{\link[=addDataPath]{addDataPath()}},
\code{\link[=addDataset]{addDataset()}},
\code{\link[=addPath]{addPath()}},
\code{\link[=addVariable]{addVariable()}},
\code{\link[=changePath]{changePath()}},
\code{\link[=convertPath]{convertPath()}},
\code{\link[=nodes]{nodes()}},
\code{\link[=paths]{paths()}},
\code{\link[=removeConstant]{removeConstant()}},
\code{\link[=removeDataPath]{removeDataPath()}},
\code{\link[=removePath]{removePath()}},
\code{\link[=removeVariable]{removeVariable()}}
}
\concept{GraphModel verbs}
