% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validateOptimizationParams}
\alias{validateOptimizationParams}
\title{Validate Optimization Parameters}
\usage{
validateOptimizationParams(schema)
}
\arguments{
\item{schema}{A validated schema list}
}
\value{
Invisibly returns TRUE. Throws error if invalid.
}
\description{
Checks that:
\itemize{
\item Fixed parameters have values (not null)
\item Free parameters have valid values or are null (will default to 0.1)
\item Free is "free" or "fixed"
}
}
\keyword{internal}
